+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
-All notable changes to the "aco4tsp" pypi package will be documented in this file.
+All notable changes to the "aco4tsp" pypi package will be documented in this file.  
+This project follows [semantic versioning](https://semver.org/).
 
 ## Unreleased - v0.0.1
 * **Initial Release**
-* **Added** a feature that enables solving the Traveling Salesman Problem using the Ant Colony Optimization approach
+* **Added** a feature that enables solving the Traveling Salesman Problem using the Ant Colony Optimization approach and plots the result afterwards